Fix dashboard nav
