--- front/src/dashboard/DashboardNavigation.module.scss
@@ -1,73 +1,0 @@
-@import "src/scss/Constants.scss";
-
-.dashboardNav {
-  background-color: $grey-lighter;
-  &CompanyTitle {
-    margin: 1rem 1rem;
-    font-weight: bold;
-    overflow-wrap: break-word;
-    // @media (min-width: 749px) {
-    //   max-width: 200px;
-    // }
-  }
-  &Chapter {
-    font-weight: bold;
-    padding: 1rem 0 0.5rem 0.5rem;
-    color: $grey-darkest;
-    font-size: 1.2rem;
-    margin: 0;
-  }
-  &Link {
-    display: block;
-    height: 30px;
-    line-height: 30px;
-    border-left: 3px solid transparent;
-    padding-left: 0.5rem;
-
-    &:hover {
-      background-color: $white;
-      border-color: $blue-dark;
-    }
-  }
-  &Indented {
-    padding-left: 1.5rem;
-  }
-  &Main {
-    margin-top: 0.5rem;
-    font-weight: bold;
-  }
-
-  &Active {
-    font-weight: bold;
-    background-color: $white;
-    border-color: $blue-dark;
-  }
-  li {
-    list-style: none;
-    a {
-      text-decoration: none;
-      color: $black;
-    }
-  }
-}
-.companySelect {
-  margin: 10px 10px;
-  h3 {
-    font-size: 1.2em;
-  }
-  p {
-    font-size: 0.9em;
-  }
-  & select {
-    width: calc(100%-1px);
-    @media (min-width: 749px) {
-      max-width: 200px;
-    }
-  }
-
-  p,
-  h3 {
-    font-weight: 400;
-    margin: 0;
-  }
-}
