+++ front/src/dashboard/DashboardTabs.scss
